UW Bot/requirements.txt
discord.py>=2.0.0
pandas
openpyxl
python-dotenv
